refactor(select-input): migrate to typescript
